pyproject.toml
[tool.poetry]
name = "retail-skypro"
version = "0.1.0"
description = ""
authors = ["DanilMirosh <miroshnichenko.da.vi@mail.ru>"]
readme = "README.md"
packages = [{include = "retail_skypro"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
envparse = "^0.2.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
